\name{ibestGUI-package}
\alias{ibestGUI-package}
\docType{package}
\title{
ibestGUI
}
\description{
mcaGUI is a Simple GUI used in the analysis of microbial communities using RGtk2 and gWidgetsRGtk.
}
\details{
  
Further information is available in the following vignettes:
\tabular{ll}{
\code{manual} \tab pmg (source, pdf)\cr
}
}
\author{
Wade Copeland with original PMG source by John Verzani
Authors: Wade K. Copeland, Vandhana Krishnan, Daniel Beck, Matt Settles, Zaid Abdo

Authors:
Wade K. Copeland, Vandhana Krishnan, Daniel Beck, Matt Settles, James Foster, Kyu-Chul Cho                                                    
Mitch Day, Roxana Hickey, Ursel M.E. Schutte, Xia Zhou, Chris Williams, Larry J. Forney, Zaid Abdo

Additional Input and Suggestions:
John Verzani - Author of Original of the PMG Source Code used as a backend for mcaGUI.


Maintainer: Wade K. Copeland <wade@kingcopeland.com>
URL: \url{http://www.ibest.uidaho.edu/ibest/index.php}
}
\keyword{ package }
