\name{pmg-undocumented}
\alias{pmg-undocumented.Rd}
\alias{kurtosis}
\alias{skewness}
\alias{plotecdf}
\alias{cluster}
\alias{richness}
\alias{diversityfun}
\alias{bootstrap}
\alias{principle_component}

\title{Undocumented, but exported, functions}
\description{
  These functions are used in the global environemt and so must be
  exported. However, they are not intended for general use.
}
\keyword{interface}% at least one, from doc/KEYWORDS
